+++ snafu/versions.py
@@ -65,7 +65,7 @@ class Version:
     def python_commands(self):
         dirpath = configs.get_cmd_dir_path()
         return [
-            dirpath.joinpath('python{}.lnk'.format(name))
+            dirpath.joinpath('python{}.exe'.format(name))
             for name in self.script_version_names
         ]
 
@@ -73,14 +73,14 @@ class Version:
     def pip_commands(self):
         dirpath = configs.get_cmd_dir_path()
         return [
-            dirpath.joinpath('pip{}.lnk'.format(name))
+            dirpath.joinpath('pip{}.exe'.format(name))
             for name in self.script_version_names
         ]
 
     @property
     def python_major_command(self):
         dirpath = configs.get_scripts_dir_path()
-        return dirpath.joinpath('python{}.lnk'.format(self.version_info[0]))
+        return dirpath.joinpath('python{}.exe'.format(self.version_info[0]))
 
     def get_installation(self):
         path = metadata.get_install_path(self.name).resolve(strict=True)