+++ HISTORY.md
@@ -11,8 +11,7 @@
 
 ### Behavioural Changes
 
-* Command aliases are now provided by as Windows shortcuts (`.lnk`). SNAFU adds ``.LNK`` to your ``PathExt`` environment variable so you can call them without the extension.
-* The `snafu` entry point command is also a shortcut now.
+* Call `snafu link --all` automatically after installing with a shimmed pip command.
 * Upgrade Python 3.4 definition to 3.4.4.
 * Improve script copy logic in `link` to speed up the process and reduce cluttering output.
 * Improve behaviour when uninstalling versions not installed by SNAFU.